wyciszenie warninga gcc
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/branches/0.6.0.1@5014 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ kadu-core/configuration_window_widgets.cpp
@@ -310,7 +310,7 @@ QString ConfigComboBox::currentItemValue()
 {
 	int index = currentItem();
 
-	if ((index < 0) || (index >= itemValues.size()))
+	if ((index < 0) || ((unsigned int)index >= itemValues.size()))
 		return QString::null;
 
 	return itemValues[index];
@@ -351,7 +351,7 @@ void ConfigComboBox::saveConfiguration()
 
 	int index = currentItem();
 
-	if ((index < 0) || (index >= itemValues.size()))
+	if ((index < 0) || ((unsigned int)index >= itemValues.size()))
 		return;
 
 	config_file.writeEntry(section, item, itemValues[currentItem()]);
